vision-master/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
opencv-python==4.10.0.84
ultralytics==8.3.0
jinja2==3.1.4
python-multipart==0.0.9
python-dotenv==1.0.0
requests==2.31.0
